Scripts/BT_TAPO/requirements.txt
tapo
python-dotenv
PyYAML